+++ capturable/src/androidMain/kotlin/dev/wonddak/capturable/CaptureController+Extension.kt
@@ -23,7 +23,7 @@
 * SOFTWARE.
 *
 */
-package dev.wonddak.capturable.util
+package dev.wonddak.capturable
 
 import android.content.ClipData
 import android.content.Context
